+++ changelog.md
@@ -8,3 +8,4 @@
 
 * `.readEntries` + `.walkEntries` avoid creating long promise chains (closes #2)
 * `.walkEntries` awaits promise from callback (closes #1)
+* `ZipFile` is not event emitter (closes #3)