+++ changelog.md
@@ -6,6 +6,7 @@
 
 ## Next
 
+* Promisify `.close` method
 * `.readEntries` + `.walkEntries` avoid creating long promise chains (closes #2)
 * `.walkEntries` awaits promise from callback (closes #1)
 * `ZipFile` is not event emitter (closes #3)